ent core API cleanup/remove dead code/add comments/put in right places
